tools/requirements.txt
# FlockBlocker Testing Framework Dependencies

# OCR Engines (for ensemble adversarial testing)
pytesseract>=0.3.10        # Tesseract OCR wrapper
easyocr>=1.7.0             # PyTorch-based multi-language OCR
paddleocr>=2.7.0           # PaddlePaddle OCR (high accuracy on plates)

# Image Processing
opencv-python>=4.8.0       # Image manipulation, perspective transforms
Pillow>=10.0.0             # Image I/O and basic processing
numpy>=1.24.0              # Array operations
scikit-image>=0.21.0       # Image quality metrics, SSIM

# ML / Adversarial
torch>=2.1.0               # PyTorch for adversarial patch generation
torchvision>=0.16.0        # Vision transforms and models

# Testing
pytest>=7.4.0              # Test runner
pytest-html>=4.0.0         # HTML test reports
pytest-cov>=4.1.0          # Coverage reporting

# Reporting
matplotlib>=3.7.0          # Visualization of results
pandas>=2.1.0              # Data analysis of test results
tabulate>=0.9.0            # Pretty-print tables
